abstract AsyncTask
