+++ app/src/main/java/com/example/hochi/nextcompanion/AsyncTaskCallbacks.java
@@ -1,0 +1,5 @@
+package com.example.hochi.nextcompanion;
+
+interface AsyncTaskCallbacks<T> {
+    void onTaskComplete(T result);
+}
